+++ android/src/main/java/com/audiobrowser/player/PlaybackStateStore.kt
@@ -191,7 +191,7 @@ class PlaybackStateStore(private val player: Player) {
           live = if (obj.has("live") && !obj.isNull("live")) obj.getBoolean("live") else null,
           imageRow = null, // Not persisted
           replayGain =
-            if (obj.has("replayGain") && !obj.isNull("replayGain")) obj.getFloat("replayGain") else null,
+            if (obj.has("replayGain") && !obj.isNull("replayGain")) obj.getDouble("replayGain") else null,
         )
       }
       .onFailure { e -> Timber.w(e, "Failed to parse persisted track JSON") }
